move ToPrettyString serde into shim layer
