+++ spark/src/main/spark-3.5/org/apache/comet/shims/CometExprShim.scala
@@ -18,10 +18,14 @@
  */
 package org.apache.comet.shims
 
-import org.apache.comet.expressions.CometEvalMode
-import org.apache.comet.serde.CommonStringExprs
+import org.apache.comet.CometSparkSessionExtensions.withInfo
+import org.apache.comet.expressions.{CometCast, CometEvalMode}
+import org.apache.comet.serde.{CommonStringExprs, Compatible, ExprOuterClass, Incompatible}
 import org.apache.comet.serde.ExprOuterClass.{BinaryOutputStyle, Expr}
+import org.apache.comet.serde.QueryPlanSerde.{binaryOutputStyle, exprToProtoInternal}
+
 import org.apache.spark.sql.catalyst.expressions._
+import org.apache.spark.sql.types.DataTypes
 
 /**
  * `CometExprShim` acts as a shim for parsing expressions from different Spark versions.
@@ -40,6 +44,44 @@ trait CometExprShim extends CommonStringExprs {
         case s: StringDecode =>
           // Right child is the encoding expression.
           stringDecode(expr, s.charset, s.bin, inputs, binding)
+
+        case expr: ToPrettyString =>
+          val child = expr.asInstanceOf[UnaryExpression].child
+          val timezoneId = expr.asInstanceOf[TimeZoneAwareExpression].timeZoneId
+
+          val castSupported = CometCast.isSupported(
+            child.dataType,
+            DataTypes.StringType,
+            timezoneId,
+            CometEvalMode.TRY)
+
+          val isCastSupported = castSupported match {
+            case Compatible(_) => true
+            case Incompatible(_) => true
+            case _ => false
+          }
+
+          if (isCastSupported) {
+            exprToProtoInternal(child, inputs, binding) match {
+              case Some(p) =>
+                val toPrettyString = ExprOuterClass.ToPrettyString
+                  .newBuilder()
+                  .setChild(p)
+                  .setTimezone(timezoneId.getOrElse("UTC"))
+                  .setBinaryOutputStyle(binaryOutputStyle)
+                  .build()
+                Some(
+                  ExprOuterClass.Expr
+                    .newBuilder()
+                    .setToPrettyString(toPrettyString)
+                    .build())
+              case _ =>
+                withInfo(expr, child)
+                None
+            }
+          } else {
+            None
+          }
 
         case _ => None
       }
